+++ exporters/otlp/testing-internal/src/main/java/io/opentelemetry/exporter/otlp/testing/internal/GrpcMetricExporterBuilderWrapper.java
@@ -8,76 +8,76 @@ package io.opentelemetry.exporter.otlp.testing.internal;
 import io.grpc.ManagedChannel;
 import io.opentelemetry.exporter.internal.retry.RetryPolicy;
 import io.opentelemetry.exporter.internal.retry.RetryUtil;
-import io.opentelemetry.exporter.otlp.logs.OtlpGrpcLogExporterBuilder;
-import io.opentelemetry.sdk.logs.data.LogData;
+import io.opentelemetry.exporter.otlp.metrics.OtlpGrpcMetricExporterBuilder;
+import io.opentelemetry.sdk.metrics.data.MetricData;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
 
-final class GrpcLogExporterWrapper implements TelemetryExporterBuilder<LogData> {
-  private final OtlpGrpcLogExporterBuilder builder;
+final class GrpcMetricExporterBuilderWrapper implements TelemetryExporterBuilder<MetricData> {
+  private final OtlpGrpcMetricExporterBuilder builder;
 
-  GrpcLogExporterWrapper(OtlpGrpcLogExporterBuilder builder) {
+  GrpcMetricExporterBuilderWrapper(OtlpGrpcMetricExporterBuilder builder) {
     this.builder = builder;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> setEndpoint(String endpoint) {
+  public TelemetryExporterBuilder<MetricData> setEndpoint(String endpoint) {
     builder.setEndpoint(endpoint);
     return this;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> setTimeout(long timeout, TimeUnit unit) {
+  public TelemetryExporterBuilder<MetricData> setTimeout(long timeout, TimeUnit unit) {
     builder.setTimeout(timeout, unit);
     return this;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> setTimeout(Duration timeout) {
+  public TelemetryExporterBuilder<MetricData> setTimeout(Duration timeout) {
     builder.setTimeout(timeout);
     return this;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> setCompression(String compression) {
+  public TelemetryExporterBuilder<MetricData> setCompression(String compression) {
     builder.setCompression(compression);
     return this;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> addHeader(String key, String value) {
+  public TelemetryExporterBuilder<MetricData> addHeader(String key, String value) {
     builder.addHeader(key, value);
     return this;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> setTrustedCertificates(byte[] certificates) {
+  public TelemetryExporterBuilder<MetricData> setTrustedCertificates(byte[] certificates) {
     builder.setTrustedCertificates(certificates);
     return this;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> setClientTls(
+  public TelemetryExporterBuilder<MetricData> setClientTls(
       byte[] privateKeyPem, byte[] certificatePem) {
     builder.setClientTls(privateKeyPem, certificatePem);
     return this;
   }
 
   @Override
-  public TelemetryExporterBuilder<LogData> setRetryPolicy(RetryPolicy retryPolicy) {
+  public TelemetryExporterBuilder<MetricData> setRetryPolicy(RetryPolicy retryPolicy) {
     RetryUtil.setRetryPolicyOnDelegate(builder, retryPolicy);
     return this;
   }
 
   @Override
   @SuppressWarnings("deprecation") // testing deprecated functionality
-  public TelemetryExporterBuilder<LogData> setChannel(ManagedChannel channel) {
+  public TelemetryExporterBuilder<MetricData> setChannel(ManagedChannel channel) {
     builder.setChannel(channel);
     return this;
   }
 
   @Override
-  public TelemetryExporter<LogData> build() {
+  public TelemetryExporter<MetricData> build() {
     return TelemetryExporter.wrap(builder.build());
   }
 }
